Simplify new custom-catalog code and flow
