+++ ynh-dev
@@ -19,7 +19,11 @@ function show_usage() {
   test [PKG]      Deploy, update and run tests for some packages
                   Tests for single modules and functions can ran with
                   e.g. ./ynh-dev test yunohost/appurl:urlavailable
-
+  catalog 
+      build       Rebuild the custom catalog
+      add         Add the custom catalog in Yunohost catalog list
+      override    Override default catalog with the custom catalog
+      reset       Reset the catalog list to Yunohost's default
 EOF
 }
 
@@ -489,7 +493,7 @@ function catalog()
 {
     assert_inside_vm
     local ACTION="$1"
-    local CUSTOM_APPS_FOLDER=${2:-"/ynh-dev/custom-apps"}
+    local CUSTOM_APPS_FOLDER=${2:-"/ynh-dev/custom-catalog"}
     local CUSTOM_CAT_PATH="${CUSTOM_APPS_FOLDER}/catalog.json"
     local CACHE_FOLDER="/var/cache/yunohost/repo"
 
@@ -498,22 +502,22 @@ function catalog()
     case "${ACTION}" in
         build)
             info "Rebuilding custom app catalog"
-            python3 -c "from catalog_builder import build; build(folder='${CUSTOM_APPS_FOLDER}')" && success "Successfully build custom catalog list in '${CUSTOM_CAT_PATH}'"
+            python3 -c "from catalog_manager import build; build(folder='${CUSTOM_APPS_FOLDER}')" && success "Successfully build custom catalog list in '${CUSTOM_CAT_PATH}'"
             ;;
         add)
             info "Injecting custom catalog in YunoHost catalog list"
             create_sym_link "${CUSTOM_CAT_PATH}" "${CACHE_FOLDER}/custom.json"
-            python3 -c "from catalog_list import add; add()" && success "Custom catalog '${CUSTOM_CAT_PATH}' added to catalog list"
+            python3 -c "from catalog_manager import add; add()" && success "Custom catalog '${CUSTOM_CAT_PATH}' added to catalog list"
             ;;
         override)
             info "Overriding default catalog by custom one"
             create_sym_link "${CUSTOM_CAT_PATH}" "${CACHE_FOLDER}/custom.json"
-            python3 -c "from catalog_list import override; override()" && success "Default catalog is now overrided by '$CUSTOM_CAT_PATH'"
+            python3 -c "from catalog_manager import override; override()" && success "Default catalog is now overrided by '$CUSTOM_CAT_PATH'"
             ;;
         reset)
             info "Reseting to YunoHost default catalog list"
             [ -e "$CACHE_FOLDER/custom.json" ] && rm "$CACHE_FOLDER/custom.json"
-            python3 -c "from catalog_list import reset; reset()" || exit 1
+            python3 -c "from catalog_manager import reset; reset()" || exit 1
             success "Returned to default"
             ;;
         *)
